use css-minimizer-webpack-plugin instead of own implementation of cssnano which should integrate better with webpack
