+++ lib/index.ts
@@ -26,6 +26,6 @@ export { default as FreshheadsDefaultStackAdapter } from './adapter/freshheads/D
 export { default as FreshheadsDefaultAdminStackAdapter } from './adapter/freshheads/DefaultAdminStackAdapter';
 export { default as FreshheadsDefaultSonataAdminStackAdapter } from './adapter/freshheads/DefaultSonataAdminStackAdapter';
 export { default as FreshheadsBabelLoaderAdapter } from './adapter/freshheads/BabelLoaderAdapter';
-export { default as FreshheadsJavascriptMinimizationAdapter } from './adapter/freshheads/JavascriptMinimizationAdapter';
+export { default as FreshheadsCssMinimizationAdapter } from './adapter/freshheads/CssMinimizationAdapter';
 export { default as FreshheadsJavascriptJQueryAdapter } from './adapter/freshheads/JavascriptJQueryAdapter';
 export { default as FreshheadsSassLoaderAdapter } from './adapter/freshheads/SassLoaderAdapter';
