+++ package.json
@@ -54,7 +54,7 @@
         "babel-loader": "^8.2.2",
         "copy-webpack-plugin": "^8.0.0",
         "css-loader": "^5.1.1",
-        "cssnano": "^4.1.10",
+        "css-minimizer-webpack-plugin": "^1.2.0",
         "fibers": "^5.0.0",
         "file-loader": "^6.2.0",
         "husky": "^5.1.3",
